+++ src/clients/rate_limited_kraken_client.rs
@@ -97,6 +97,20 @@ where
         }
     }
 
+    fn new_with_tracing(
+        secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>>,
+        nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>>,
+        trace_inbound: bool,
+    ) -> Self {
+        RateLimitedKrakenClient {
+            core_client: C::new_with_tracing(secrets_provider, nonce_provider, trace_inbound),
+            private_rate_limiter: Self::get_private_rate_limiter(VerificationTier::Intermediate),
+            public_rate_limiter: Self::get_public_rate_limiter(),
+            trading_rate_limiter: KrakenTradingRateLimiter::new(VerificationTier::Intermediate),
+            pair_rate_limiter: KeyedRateLimiter::new(),
+        }
+    }
+
     async fn set_user_agent(&mut self, user_agent: String) {
         self.core_client.set_user_agent(user_agent).await;
     }