+++ tests/resources/test_client/test_client_impl.rs
@@ -74,6 +74,14 @@ impl KrakenClient for TestClient {
         Self { current_id: 0 }
     }
 
+    fn new_with_tracing(
+        _secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>>,
+        _nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>>,
+        _trace_inbound: bool,
+    ) -> Self {
+        Self { current_id: 0 }
+    }
+
     async fn set_user_agent(&mut self, _user_agent: String) {}
 
     async fn get_server_time(&mut self) -> Result<ResultErrorResponse<SystemTime>, ClientError> {